// frontend/static/enums/WebSocketEvents.ts
export enum WebSocketEvents {
    RegisterForCanvas = "RegisterForCanvas",
    RegisteredForCanvas = "RegisteredForCanvas",
    CanvasEvent = "CanvasEvent",
    CreatedClientId = "CreatedClientId",
    SessionID = "SessionID",
    CreateCanvas = "CreateCanvas",
    CanvasCreated = "CanvasCreated",
    DeregisterForCanvas = "DeregisterForCanvas",
    CanvasChangedEvent = "CanvasChangedEvent",
    GetCanvasEvents = "GetCanvasEvents",
    GetCanvasEventsResponse = "GetCanvasEventsResponse",
}